Config package refactor
