+++ shortcut_composer/config_system/common_utils/__init__.py
@@ -1,0 +1,4 @@
+from .api_krita import Krita
+from .save_location import SaveLocation
+
+__all__ = ["Krita", "SaveLocation"]
