+++ shortcut_composer/config_system/field_base_impl/__init__.py
@@ -1,7 +1,14 @@
 # SPDX-FileCopyrightText: © 2022-2023 Wojciech Trybus <wojtryb@gmail.com>
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from .non_list_field import NonListField
+from .list_field import ListField
 from .dual_field import DualField
 from .field_with_editable_default import FieldWithEditableDefault
 
-__all__ = ["DualField", "FieldWithEditableDefault"]
+__all__ = [
+    "NonListField",
+    "ListField",
+    "DualField",
+    "FieldWithEditableDefault"
+]
