+++ shortcut_composer/config_system/field_base_impl/common_utils/__init__.py
@@ -1,0 +1,3 @@
+from .parsers import dispatch_parser, Parser
+
+__all__ = ["dispatch_parser", "Parser"]
